Adds support for no video/audio
This codebase was originally designed to fully integrate OMXPlayer and DMX. However, in some instances running OMXPlayer is not necessary.

This change introduces a Mock class for OMXPlayer which effectivelly emulates the interface of the OMXPlayer class but without any actual video output.

This Mock class now automatically instatiates if a media file is not found (or the media file variable is set to None).